+++ edge/deleteMergedBranches.ts
@@ -5,6 +5,7 @@ import * as Util from './Util'
 import * as Git from './Git'
 import { fetchInternal } from './fetch'
 import Log from './Log'
+import { track } from './Amplitude'
 
 let syncingStatusBar: vscode.StatusBarItem