+++ edge/pullRequest.ts
@@ -3,6 +3,7 @@ import * as open from 'open'
 import * as Util from './Util'
 import * as Git from './Git'
 import push from './push'
+import { track } from './Amplitude'
 
 export default async function () {
 	const workspace = await Util.getCurrentWorkspace()