+++ edge/sync.ts
@@ -4,6 +4,7 @@ import * as vscode from 'vscode'
 import * as Util from './Util'
 import * as Git from './Git'
 import { getMergedBranchNames } from './deleteMergedBranches'
+import { track } from './Amplitude'
 
 export default async function () {
 	track('sync')